readyup|refactor: Rename src/packages to installed-packages
Renames `packages/readyup/src/packages/` to `installed-packages/`, whose three modules locate and interrogate npm packages installed in the project under check. The old name repeated the repo's workspace directory, so both the path and a specifier like `../packages/resolvePackageRoot.ts` read two ways.

`packagesRun.wiring.unit.test.ts` moves to `run/__tests__/`, beside the `runCommand.ts` it exercises. It had no subject in the directory it sat in: `packages` names the `--packages` flag as readily as the directory's own code, which is how a test of that flag came to sit there.
