+++ packages/readyup/src/check-utils/discoverKitPackages.ts
@@ -2,9 +2,9 @@ import { readFileSync } from 'node:fs';
 import path from 'node:path';
 import process from 'node:process';
 
+import { resolvePackageRoot } from '../installed-packages/resolvePackageRoot.ts';
 import { KITS_DIR } from '../kits/kitsDir.ts';
 import { enumerateKits } from '../list/enumerateKits.ts';
-import { resolvePackageRoot } from '../packages/resolvePackageRoot.ts';
 import { isRecord } from '../portable/isRecord.ts';
 
 /** Manifest fields a kit-carrying dependency can be declared in. */
