+++ packages/readyup/src/installed-packages/__tests__/expandConfiguredPackages.unit.test.ts
