+++ packages/readyup/src/installed-packages/__tests__/isPackageInstalled.unit.test.ts
