+++ packages/readyup/src/installed-packages/__tests__/resolvePackageRoot.unit.test.ts
