+++ packages/readyup/src/installed-packages/expandConfiguredPackages.ts
