+++ packages/readyup/src/installed-packages/isPackageInstalled.ts
