+++ packages/readyup/src/installed-packages/resolvePackageRoot.ts
