+++ packages/readyup/src/list/listCommand.ts
@@ -11,6 +11,8 @@ import { DEFAULT_CONFIG, loadConfig } from '../config/loadConfig.ts';
 import { extractHint } from '../errors/error-handling.ts';
 import { translateParseArgsError } from '../errors/parse-args-error.ts';
 import { configError, kitLoadError, usageError } from '../errors/RdyError.ts';
+import { expandConfiguredPackages, type PackageKit } from '../installed-packages/expandConfiguredPackages.ts';
+import { resolvePackageRoot } from '../installed-packages/resolvePackageRoot.ts';
 import { KITS_DIR, resolveHomeDir } from '../kits/kitsDir.ts';
 import type { DirectorySource, GlobalSource, LocalSource, NpmSource } from '../kits/parseFromValue.ts';
 import { parseFromValue } from '../kits/parseFromValue.ts';
@@ -20,8 +22,6 @@ import { DEFAULT_MANIFEST_PATH } from '../manifest/manifestPath.ts';
 import type { RdyManifest, RdyManifestKit } from '../manifest/manifestSchema.ts';
 import { ManifestNotFoundError, readManifest } from '../manifest/readManifest.ts';
 import { writeHuman } from '../output/writeHuman.ts';
-import { expandConfiguredPackages, type PackageKit } from '../packages/expandConfiguredPackages.ts';
-import { resolvePackageRoot } from '../packages/resolvePackageRoot.ts';
 import { isSkippableFilesystemError } from '../portable/isSkippableFilesystemError.ts';
 import type { KitProject } from '../projects/discoverKitProjects.ts';
 import { discoverKitProjects } from '../projects/discoverKitProjects.ts';
