+++ packages/readyup/src/run/__tests__/packagesRun.wiring.unit.test.ts
@@ -5,9 +5,9 @@ import process from 'node:process';
 
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { type ResolvedKitEntry, resolveKitSources, runCommand } from '../../run/runCommand.ts';
 import { ReportSchema } from '../../schemas/reportSchema.ts';
 import { captureRdyError } from '../../test-utils/captureRdyError.ts';
+import { type ResolvedKitEntry, resolveKitSources, runCommand } from '../runCommand.ts';
 
 /**
  * Joins `--packages` to the kits an installed package publishes, against a real fixture project. The
